backend/requirement.txt
fastapi
uvicorn
pydantic
groq
python-dotenv
qdrant-client[fastembed]
langchain-text-splitters
